Server/requirements.txt
Flask==0.12.2
pillow==5.4.1
requests
bs4
keras
